requirements.txt
django==1.11.3
psycopg2==2.7.1
psycopg2-binary==2.7.4
django_heroku==0.3.1
gunicorn==19.7.1
reportlab
xhtml2pdf>=0.2b1
django-easy-pdf